Split upload tests out of aggregator.rs
